books/pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
strawberry-graphql = "^0.31.1"
starlette = "^0.13.8"
uvicorn = "^0.11.8"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
